Version 0.3.1
Re-designed options frame
Changed "highest available" video format to "default"

Fixed password bug
Fixed "Download subtitle file by languages" bug

Removed Options
	Rate Limit
Added Options
	Embed Subtitles to mp4
	Video Password (Vimeo etc..)

Added version check for settings file

+++ youtube_dl_gui/Utils.py
@@ -28,7 +28,8 @@ def have_dash_audio(audio):
   return audio != "NO SOUND"
   
 def remove_file(filename):
-  os.remove(filename)
+  if file_exist(filename):
+    os.remove(filename)
 
 def get_path_seperator():
   if os.name == 'nt':
